+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-app",
+  "name": "typescriptapp",
   "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
@@ -9,10 +9,17 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "expo": "~50.0.4",
+    "@gluestack-style/react": "^1.0.45",
+    "@gluestack-ui/config": "^1.1.2",
+    "@gluestack-ui/themed": "^1.1.3",
+    "@tanstack/react-query": "^5.18.1",
+    "expo": "~50.0.6",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
-    "react-native": "0.73.2"
+    "react-native": "0.73.2",
+    "react-native-svg": "^13.4.0",
+    "react-native-swiper": "^1.6.0",
+    "@expo/metro-runtime": "~3.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
